modify how plantuml injects itself in the kramdown process
This commit changes how the injection happens from plain monkey
patching to using a module and prepending it into the class hierarchy.

This technique fixes compatibility with middleman. Middleman creates
a subclass of Kramdown::Converter::Html, and this was making the
method-overriding technique fail, while this one works.

To use in middleman, just require "kramdown-plantuml" in config.rb